Refactor source

+++ super-duper-driver/src/main/java/com/udacity/jwdnd/course1/superduperdriver/controller/CredentialController.java
@@ -1,7 +1,9 @@
 package com.udacity.jwdnd.course1.superduperdriver.controller;
 
+import com.udacity.jwdnd.course1.superduperdriver.common.Constants;
 import com.udacity.jwdnd.course1.superduperdriver.model.entities.Credential;
 import com.udacity.jwdnd.course1.superduperdriver.service.CredentialService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ModelAttribute;
@@ -9,43 +11,56 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import java.security.Principal;
+
 @Controller
 @RequestMapping("/credential")
 public class CredentialController {
 
-    private final CredentialService credentialService;
+    @Autowired
+    CredentialService credentialService;
 
-    public CredentialController(final CredentialService credentialService) {
-        this.credentialService = credentialService;
-    }
-
-    @PostMapping()
-    public String addCredential(@ModelAttribute Credential credential, Model model) {
-        if (credentialService.addCredential(credential) > 0) {
-            model.addAttribute("addCredentialSuccess", "Insert credential success.");
-        } else {
-            model.addAttribute("addCredentialFail", "Insert credential failure.");
+    @PostMapping("/add")
+    public String addCredential(@ModelAttribute Credential credential, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (credentialService.addCredential(credential) > 0) {
+            model.addAttribute("message", "Insert credential success.");
+        } else {
+            model.addAttribute("error", "Insert credential failed.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 
     @PostMapping("/update")
-    public String updateCredential(@ModelAttribute Credential credential, Model model) {
-        if (credentialService.updateCredential(credential) > 0) {
-            model.addAttribute("updateNoteSuccess", "Insert note success.");
-        } else {
-            model.addAttribute("updateNoteFail", "Insert note failure.");
+    public String updateCredential(@ModelAttribute Credential credential, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (credentialService.updateCredential(credential) > 0) {
+            model.addAttribute("message", "Update credential success.");
+        } else {
+            model.addAttribute("error", "Update credential failed.");
+        }
+        return Constants.REDIRECT_HOME;
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteCredential(@PathVariable Integer id, Model model) {
-        if (credentialService.deleteNote(id) > 0) {
-            model.addAttribute("updateNoteSuccess", "Insert note success.");
-        } else {
-            model.addAttribute("updateNoteFail", "Insert note failure.");
+    public String deleteCredential(@PathVariable Integer id, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (credentialService.deleteNote(id) > 0) {
+            model.addAttribute("message", "Delete credential success.");
+        } else {
+            model.addAttribute("error", "Delete note failed.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 }
