+++ super-duper-driver/src/main/java/com/udacity/jwdnd/course1/superduperdriver/controller/FileController.java
@@ -1,7 +1,11 @@
 package com.udacity.jwdnd.course1.superduperdriver.controller;
 
-import com.udacity.jwdnd.course1.superduperdriver.model.entities.FileEntity;
+import com.udacity.jwdnd.course1.superduperdriver.common.Constants;
+import com.udacity.jwdnd.course1.superduperdriver.config.FileUploadProperties;
+import com.udacity.jwdnd.course1.superduperdriver.model.entities.File;
+import com.udacity.jwdnd.course1.superduperdriver.service.AuthenticationService;
 import com.udacity.jwdnd.course1.superduperdriver.service.FileService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.core.io.Resource;
 import org.springframework.http.HttpHeaders;
@@ -16,30 +20,48 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.security.Principal;
+
 @Controller
 @RequestMapping("/file")
 public class FileController {
 
-    private final FileService fileService;
+    @Autowired
+    FileService fileService;
 
-    public FileController(final FileService fileService) {
-        this.fileService = fileService;
-    }
+    @Autowired
+    AuthenticationService authenticationService;
+
+    @Autowired
+    FileUploadProperties fileUploadProperties;
 
     @PostMapping("/upload")
-    public String uploadFile(@RequestParam("fileUpload") MultipartFile file, Model model) {
-        if (fileService.addFile(file) > 0) {
-            model.addAttribute("addFileSuccess", "Upload file success.");
-        } else {
-            model.addAttribute("addFileFail", "Upload file failure.");
+    public String uploadFile(@RequestParam("fileUpload") MultipartFile file, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        int userId = authenticationService.getUserId();
+        long maxFileSize = convertFileSize(fileUploadProperties.getMaxFileSize());
+        if (file.isEmpty()) {
+            model.addAttribute("error", "Please select a file to upload.");
+        } else if (file.getSize() > maxFileSize) {
+            model.addAttribute("error", "Please upload file under 10MB in size.");
+        } else if (fileService.checkExistFileName(file.getOriginalFilename(), userId)) {
+            model.addAttribute("error", "File name is already exists.");
+        } else if (fileService.addFile(file) < 1) {
+            model.addAttribute("error", "Upload file failed.");
+        } else {
+            model.addAttribute("message", "Upload file success.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 
     @GetMapping("/download/{id}")
     public ResponseEntity<Resource> downloadFile(@PathVariable Integer id) throws Exception {
         try {
-            FileEntity file = fileService.getFileDetail(id);
+            File file = fileService.getFileDetail(id);
             return ResponseEntity.ok()
                                  .contentType(MediaType.parseMediaType(file.getContentType()))
                                  .header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + file.getFileName() + "\"")
@@ -50,12 +72,26 @@ public class FileController {
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteFile(@PathVariable Integer id, Model model) {
+    public String deleteFile(@PathVariable Integer id, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
+        }
+
         if (fileService.deleteFile(id) > 0) {
             model.addAttribute("updateNoteSuccess", "Insert note success.");
         } else {
             model.addAttribute("updateNoteFail", "Insert note failure.");
         }
-        return "redirect:/home";
+        return Constants.REDIRECT_HOME;
+    }
+
+    private long convertFileSize(String fileSize) {
+        String sizeValue = fileSize.replaceAll("[^\\d.]", "");
+        if (fileSize.endsWith("MB")) {
+            return (long) (Double.parseDouble(sizeValue) * 1024 * 1024);
+        } else if (fileSize.endsWith("KB")) {
+            return (long) (Double.parseDouble(sizeValue) * 1024);
+        }
+        return Long.parseLong(sizeValue);
     }
 }
