+++ super-duper-driver/src/main/java/com/udacity/jwdnd/course1/superduperdriver/controller/NoteController.java
@@ -1,53 +1,67 @@
 package com.udacity.jwdnd.course1.superduperdriver.controller;
 
-import com.udacity.jwdnd.course1.cloudstorage.model.entities.Note;
-import com.udacity.jwdnd.course1.cloudstorage.services.NoteService;
+import com.udacity.jwdnd.course1.superduperdriver.common.Constants;
+import com.udacity.jwdnd.course1.superduperdriver.model.entities.Note;
+import com.udacity.jwdnd.course1.superduperdriver.service.NoteService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import java.security.Principal;
 
 @Controller
 @RequestMapping("/note")
 public class NoteController {
 
-    private final NoteService noteService;
+    @Autowired
+    NoteService noteService;
 
-    public NoteController(final NoteService noteService) {
-        this.noteService = noteService;
-    }
-
-    @PostMapping()
-    public String addNote(@ModelAttribute Note note, Model model) {
-        if (noteService.addNote(note) > 0) {
-            model.addAttribute("addNoteSuccess", "Insert note success.");
-        } else {
-            model.addAttribute("addNoteFail", "Insert note failure.");
+    @PostMapping("/add")
+    public String addNote(@ModelAttribute Note note, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (noteService.addNote(note) > 0) {
+            model.addAttribute("message", "Insert note success.");
+        } else {
+            model.addAttribute("error", "Insert note failed.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 
     @PostMapping("/update")
-    public String updateNote(@ModelAttribute Note note, Model model) {
-        if (noteService.updateNote(note) > 0) {
-            model.addAttribute("updateNoteSuccess", "Insert note success.");
-        } else {
-            model.addAttribute("updateNoteFail", "Insert note failure.");
+    public String updateNote(@ModelAttribute Note note, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (noteService.updateNote(note) > 0) {
+            model.addAttribute("message", "Insert note success.");
+        } else {
+            model.addAttribute("error", "Insert note failure.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 
     @PostMapping("/delete/{id}")
-    public String deleteNote(@PathVariable Integer id, Model model) {
-        if (noteService.deleteNote(id) > 0) {
-            model.addAttribute("updateNoteSuccess", "Insert note success.");
-        } else {
-            model.addAttribute("updateNoteFail", "Insert note failure.");
+    public String deleteNote(@PathVariable Integer id, Model model, Principal principal) {
+        if (principal == null) {
+            return Constants.REDIRECT_LOGIN;
         }
-        return "redirect:/home";
+
+        if (noteService.deleteNote(id) > 0) {
+            model.addAttribute("message", "Insert note success.");
+        } else {
+            model.addAttribute("error", "Insert note failure.");
+        }
+
+        return Constants.REDIRECT_HOME;
     }
 }
